job-agent/requirements.txt
flask
requests
PyPDF2
python-dotenv
python-docx
reportlab
pytest
